Simple server handle multiple concurrent requests from same client
