+++ tests/conftest.py
@@ -2,3 +2,4 @@ from __future__ import annotations
 
 from testing.fixture import exchange
 from testing.fixture import launcher
+from testing.fixture import simple_exchange_server
